rename: index 작업
