+++ src/components/index.ts
@@ -1,0 +1,7 @@
+export { default as SquareLayout } from './SquareLayout.vue';
+export { default as Navigation } from './Navigation.vue';
+export { default as BasicButton } from './BasicButton.vue';
+export { default as CandidateBox } from './CandidateBox.vue';
+export { default as CandidateList } from './CandidateList.vue';
+export { default as BarChart } from './BarChart.vue';
+export { default as Pagination } from './Pagination.vue';
