+++ tsconfig.json
@@ -23,6 +23,7 @@
     /* Alias */
     "baseUrl": ".",
     "paths": {
+      "@components": ["src/components/index"],
       "@components/*": ["src/components/*"],
       "@pages/*": ["src/pages/*"],
       "@routes/*": ["src/routes/*"],
